+++ src/js/features/camera-scanner/toggle-torch-button-status.js
