+++ src/js/features/file-scanner/file-scanner.js
