+++ src/js/features/scan-results/create-result.js
