+++ src/js/features/scan-results/scan-result.js
@@ -1,6 +1,6 @@
 import { isWebShareSupported } from '@georapbox/web-share-element/dist/is-web-share-supported.js';
-import { getSettings } from '../services/storage.js';
-import { dateTimeFormatter } from '../utils/datetime-formatter.js';
+import { getSettings } from '../../shared/storage/storage.js';
+import { dateTimeFormatter } from '../../shared/utils/datetime-formatter.js';
 
 const styles = /* css */ `
   :host {
