+++ src/js/shared/components/clipboard-copy.js
