+++ src/js/shared/feedback/vibrate.js
